--- codebus.apigateway.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0858C000-AC4C-4E96-B1D4-8A7331A88FD5}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "codebus.apigateway.portal", "src\codebus.apigateway.portal\codebus.apigateway.portal.csproj", "{DD931BD3-2545-445A-81FE-022E8F889380}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "codebus.apigateway.core", "src\codebus.apigateway.core\codebus.apigateway.core.csproj", "{C424A6A7-BE40-437A-A8F4-398A87F4A6C0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DD931BD3-2545-445A-81FE-022E8F889380}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DD931BD3-2545-445A-81FE-022E8F889380}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DD931BD3-2545-445A-81FE-022E8F889380}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DD931BD3-2545-445A-81FE-022E8F889380}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C424A6A7-BE40-437A-A8F4-398A87F4A6C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C424A6A7-BE40-437A-A8F4-398A87F4A6C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C424A6A7-BE40-437A-A8F4-398A87F4A6C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C424A6A7-BE40-437A-A8F4-398A87F4A6C0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{DD931BD3-2545-445A-81FE-022E8F889380} = {0858C000-AC4C-4E96-B1D4-8A7331A88FD5}
-		{C424A6A7-BE40-437A-A8F4-398A87F4A6C0} = {0858C000-AC4C-4E96-B1D4-8A7331A88FD5}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6FF52B77-4CA6-46C2-8017-CA51381233C7}
-	EndGlobalSection
-EndGlobal
