+++ src/ApiGateway.Core/ApiGateway.Core.csproj
