+++ src/ApiGateway.Core/Common/Extensions.cs
@@ -1,0 +1,14 @@
+﻿using ApiGateway.Core.Model;
+
+namespace ApiGateway.Core.Common
+{
+    public static class Extensions
+    {
+        public static AjaxResult ToAjaxResult(this OperationResult result)
+        {
+            if (result.Succeeded)
+                return AjaxResult.Success(result.Message, result.Data);
+            return AjaxResult.Error(result.Message, result.Data);
+        }
+    }
+}
