+++ src/ApiGateway.Core/Model/GatewayConfiguration.cs
@@ -1,0 +1,9 @@
+﻿namespace ApiGateway.Core.Model
+{
+    public class GatewayConfiguration
+    {
+        public bool AutoUpdate { get; set; } = false;
+
+        public int UpdateInterval { get; set; }
+    }
+}
