+++ src/ApiGateway.Core/OcelotAddin/GatewayMiddleware.cs
@@ -1,0 +1,168 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Ocelot.Configuration;
+using Ocelot.Middleware;
+using System;
+using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
+using Ocelot.Configuration.Creator;
+using Ocelot.Configuration.Repository;
+using Ocelot.DependencyInjection;
+using Ocelot.Configuration.Setter;
+using Ocelot.Responses;
+using System.Linq;
+using Microsoft.AspNetCore.Hosting;
+using Ocelot.Logging;
+using System.Diagnostics;
+
+namespace ApiGateway.Core.OcelotAddin
+{
+    public static class GatewayMiddleware
+    {
+        public static async Task<IApplicationBuilder> UseGateway(this IApplicationBuilder builder)
+        {
+            await builder.UseGateway(new OcelotPipelineConfiguration());
+            return builder;
+        }
+
+        public static async Task<IApplicationBuilder> UseGateway(this IApplicationBuilder builder, Action<OcelotPipelineConfiguration> pipelineConfiguration)
+        {
+            var config = new OcelotPipelineConfiguration();
+            pipelineConfiguration?.Invoke(config);
+            return await builder.UseGateway(config);
+        }
+
+        public static async Task<IApplicationBuilder> UseGateway(this IApplicationBuilder builder, OcelotPipelineConfiguration pipelineConfiguration)
+        {
+            await CreateConfiguration(builder);
+
+            ConfigureDiagnosticListener(builder);
+
+            return CreateOcelotPipeline(builder, pipelineConfiguration);
+        }
+
+        public static Task<IApplicationBuilder> UseOcelot(this IApplicationBuilder app, Action<IApplicationBuilder, OcelotPipelineConfiguration> builderAction)
+            => UseOcelot(app, builderAction, new OcelotPipelineConfiguration());
+
+        public static async Task<IApplicationBuilder> UseOcelot(this IApplicationBuilder app, Action<IApplicationBuilder, OcelotPipelineConfiguration> builderAction, OcelotPipelineConfiguration configuration)
+        {
+            await CreateConfiguration(app);
+
+            ConfigureDiagnosticListener(app);
+
+            builderAction?.Invoke(app, configuration ?? new OcelotPipelineConfiguration());
+
+            app.Properties["analysis.NextMiddlewareName"] = "TransitionToOcelotMiddleware";
+
+            return app;
+        }
+
+        private static IApplicationBuilder CreateOcelotPipeline(IApplicationBuilder builder, OcelotPipelineConfiguration pipelineConfiguration)
+        {
+            builder.BuildOcelotPipeline(pipelineConfiguration);
+
+            /*
+            inject first delegate into first piece of asp.net middleware..maybe not like this
+            then because we are updating the http context in ocelot it comes out correct for
+            rest of asp.net..
+            */
+
+            builder.Properties["analysis.NextMiddlewareName"] = "TransitionToOcelotMiddleware";
+
+            return builder;
+        }
+
+        private static async Task<IInternalConfiguration> CreateConfiguration(IApplicationBuilder builder)
+        {
+            // make configuration from file system?
+            // earlier user needed to add ocelot files in startup configuration stuff, asp.net will map it to this
+            // var fileConfig = builder.ApplicationServices.GetService<IOptionsMonitor<FileConfiguration>>();
+            var fileConfig = await builder.ApplicationServices.GetService<IFileConfigurationRepository>().Get();
+
+            // now create the config
+            var internalConfigCreator = builder.ApplicationServices.GetService<IInternalConfigurationCreator>();
+            var internalConfig = await internalConfigCreator.Create(fileConfig.Data);
+
+            //Configuration error, throw error message
+            if (internalConfig.IsError)
+            {
+                ThrowToStopOcelotStarting(internalConfig);
+            }
+
+            // now save it in memory
+            var internalConfigRepo = builder.ApplicationServices.GetService<IInternalConfigurationRepository>();
+            internalConfigRepo.AddOrReplace(internalConfig.Data);
+
+            //fileConfig.OnChange(async (config) =>
+            //{
+            //    var newInternalConfig = await internalConfigCreator.Create(config);
+            //    internalConfigRepo.AddOrReplace(newInternalConfig.Data);
+            //});
+
+            var adminPath = builder.ApplicationServices.GetService<IAdministrationPath>();
+
+            var configurations = builder.ApplicationServices.GetServices<OcelotMiddlewareConfigurationDelegate>();
+
+            // Todo - this has just been added for consul so far...will there be an ordering problem in the future? Should refactor all config into this pattern?
+            foreach (var configuration in configurations)
+            {
+                await configuration(builder);
+            }
+
+            if (AdministrationApiInUse(adminPath))
+            {
+                //We have to make sure the file config is set for the ocelot.env.json and ocelot.json so that if we pull it from the
+                //admin api it works...boy this is getting a spit spags boll.
+                var fileConfigSetter = builder.ApplicationServices.GetService<IFileConfigurationSetter>();
+
+                // await SetFileConfig(fileConfigSetter, fileConfig);
+            }
+
+            return GetOcelotConfigAndReturn(internalConfigRepo);
+        }
+
+        private static bool AdministrationApiInUse(IAdministrationPath adminPath)
+        {
+            return adminPath != null;
+        }
+
+        //private static async Task SetFileConfig(IFileConfigurationSetter fileConfigSetter, IOptionsMonitor<FileConfiguration> fileConfig)
+        //{
+        //    var response = await fileConfigSetter.Set(fileConfig.CurrentValue);
+
+        //    if (IsError(response))
+        //    {
+        //        ThrowToStopOcelotStarting(response);
+        //    }
+        //}
+
+        private static bool IsError(Response response)
+        {
+            return response == null || response.IsError;
+        }
+
+        private static IInternalConfiguration GetOcelotConfigAndReturn(IInternalConfigurationRepository provider)
+        {
+            var ocelotConfiguration = provider.Get();
+
+            if (ocelotConfiguration?.Data == null || ocelotConfiguration.IsError)
+            {
+                ThrowToStopOcelotStarting(ocelotConfiguration);
+            }
+
+            return ocelotConfiguration.Data;
+        }
+
+        private static void ThrowToStopOcelotStarting(Response config)
+        {
+            throw new Exception($"Unable to start Ocelot, errors are: {string.Join(",", config.Errors.Select(x => x.ToString()))}");
+        }
+
+        private static void ConfigureDiagnosticListener(IApplicationBuilder builder)
+        {
+            var env = builder.ApplicationServices.GetService<IWebHostEnvironment>();
+            var listener = builder.ApplicationServices.GetService<OcelotDiagnosticListener>();
+            var diagnosticListener = builder.ApplicationServices.GetService<DiagnosticListener>();
+            diagnosticListener.SubscribeWithAdapter(listener);
+        }
+    }
+}
