+++ src/ApiGateway.Portal/appsettings.Development.json
