--- src/codebus.apigateway.core/DbRepository/GatewayDbContext.cs
@@ -1,23 +1,0 @@
-﻿using codebus.apigateway.core.DbRepository.Entities;
-using Microsoft.EntityFrameworkCore;
-
-namespace codebus.apigateway.core.DbRepository
-{
-    public class GatewayDbContext : DbContext
-    {
-        public GatewayDbContext(DbContextOptions<GatewayDbContext> options) : base(options)
-        {
-        }
-
-        public DbSet<GlobalConfiguration> GlobalConfiguration { get; set; }
-
-        public DbSet<ReRoute> ReRoute { get; set; }
-
-        public DbSet<Aggregates> Aggregates { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<GlobalConfiguration>().HasMany(x => x.ReRoutes).WithOne(x => x.GlobalConfiguration).HasForeignKey(x => x.GlobalConfigurationId);
-        }
-    }
-}
