--- src/codebus.apigateway.core/Model/GatewayConfiguration.cs
@@ -1,9 +1,0 @@
-﻿namespace codebus.apigateway.core.Model
-{
-    public class GatewayConfiguration
-    {
-        public bool AutoUpdate { get; set; } = false;
-
-        public int UpdateInterval { get; set; }
-    }
-}
