--- src/codebus.apigateway.portal/Properties/launchSettings.json
@@ -1,19 +1,0 @@
-{
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:54068",
-      "sslPort": 0
-    }
-  },
-  "profiles": {
-    "codebus.apigateway.portal": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "http://localhost:5000"
-    }
-  }
-}
